+++ api/src/main/java/team/emptyte/playersettings/api/user/User.java
@@ -2,7 +2,6 @@ package team.emptyte.playersettings.api.user;
 
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
-
 import team.emptyte.playersettings.api.storage.model.Model;
 
 import java.util.HashMap;
@@ -14,7 +13,7 @@ public class User implements Model {
 
     private final String id;
 
-    private Map<String, Byte> settings;
+    private Map<String, Integer> settings;
 
     public User(String id) {
         this.id = id;
@@ -31,15 +30,15 @@ public class User implements Model {
         return Bukkit.getPlayer(UUID.fromString(id));
     }
 
-    public Map<String, Byte> getSettings() {
+    public Map<String, Integer> getSettings() {
         return settings;
     }
 
-    public byte getSettingStatus(String id) {
+    public Integer getSettingStatus(String id) {
         return settings.get(id);
     }
 
-    public void updateSettingStatus(String id, byte status) {
+    public void updateSettingStatus(String id, Integer status) {
         settings.put(id, status);
     }
 
@@ -47,7 +46,8 @@ public class User implements Model {
         settings.remove(id);
     }
 
-    public void setSettings(Map<String, Byte> settings) {
+    public void setSettings(Map<String, Integer> settings) {
         this.settings = settings;
     }
+
 }
